requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
tensorflow 
pillow